mcp/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
openai  # For DeepSeek API compatibility
python-dotenv  # For environment variables
aiohttp  # For async HTTP calls
typing-extensions  # For enhanced typing
pydantic-settings  # For configuration management 